// src/frontend/src/backend.d.ts
export interface AdvancedStats {
    maxHP: bigint;
    currentHP: bigint;
    critChance: bigint;
    critPower: bigint;
}
export interface StatsUpdate {
    vitIncrease: bigint;
    strIncrease: bigint;
    intIncrease: bigint;
    dexIncrease: bigint;
}
export interface Character {
    xp: bigint;
    status: CharacterStatus;
    totalStatPointsEarned: bigint;
    equippedAbilities: Array<Ability>;
    class: string;
    name: string;
    season: bigint;
    level: bigint;
    baseStats: BaseStats;
    advancedStats: AdvancedStats;
    classTier: bigint;
    realm: Realm;
    totalStatPointsSpent: bigint;
}
// ...
export interface Ability {
    element: string;
    name: string;
    type: string;
    description: string;
    power: bigint;
}
export interface BaseStats {
    dex: bigint;
    int: bigint;
    str: bigint;
    vit: bigint;
}
export interface UserProfile {
    username: string;
}
export enum CharacterCreationError {
    noPermission = "noPermission",
    alreadyExists = "alreadyExists",
    limitReached = "limitReached"
}
export enum CharacterStatus {
    Dead = "Dead",
    Alive = "Alive"
}
export enum ItemType {
    Weapon = "Weapon",
    Armor = "Armor",
    Trinket = "Trinket"
}
export enum Rarity {
    Rare = "Rare",
    Uncommon = "Uncommon",
    Legendary = "Legendary",
    Common = "Common"
}
export enum Realm {
    Hardcore = "Hardcore",
    Softcore = "Softcore"
}